fix: minor fixes and refactoring NER (#539)
* feat: add train from scratch

* feat: add warning to train_from_scratch

* fix: no cudnn rnn fixed mask

* fix: session config added

* fix: warning to log

* fix: pass learning rate to default optimizer

* fix: l2 in conv layer

* fix: ner dataset reader

* feat: add returning loss to the train on batch

* doc: add model descriptions to the ner doc

* chore: move ner metrics to metrics module

* fix: remove train from scratch

* fix: add dataset names

* chore: log.info -> log.warning for warning
