+++ deeppavlov/configs/ner/ner_conll2003_pos.json
@@ -2,6 +2,7 @@
   "dataset_reader": {
     "name": "conll2003_reader",
     "data_path": "conll2003/",
+    "dataset_name": "conll2003",
     "provide_pos": true
   },
   "dataset_iterator": {
