+++ deeppavlov/dataset_readers/conll2003_reader.py
@@ -7,21 +7,22 @@ from deeppavlov.core.common.registry import register
 
 @register('conll2003_reader')
 class Conll2003DatasetReader(DatasetReader):
-    """Class to read training datasets in CONLL2003 format"""
-    def download_conll(self, dir_path):
-        download_decompress('http://files.deeppavlov.ai/deeppavlov_data/conll2003_v2.tar.gz', dir_path)
+    """Class to read training datasets in CoNLL-2003 format"""
 
-    def read(self, dir_path: str, dataset_name='conll2003', provide_pos=False):
+    def read(self, dir_path: str, dataset_name=None, provide_pos=False):
         self.provide_pos = provide_pos
         dir_path = Path(dir_path)
         files = list(dir_path.glob('*.txt'))
         if 'train.txt' not in {file_path.name for file_path in files}:
             if dataset_name == 'conll2003':
-                dir_path.mkdir(exist_ok=True, parents=True)
-                self.download_conll(dir_path)
-                files = list(dir_path.glob('*.txt'))
+                url = 'http://files.deeppavlov.ai/deeppavlov_data/conll2003_v2.tar.gz'
+            elif dataset_name == 'collection_rus':
+                url = 'http://files.deeppavlov.ai/deeppavlov_data/collection5.tar.gz'
             else:
                 raise RuntimeError('train.txt not found in "{}"'.format(dir_path))
+            dir_path.mkdir(exist_ok=True, parents=True)
+            download_decompress(url, dir_path)
+            files = list(dir_path.glob('*.txt'))
         dataset = {}
         for file_name in files:
             name = file_name.with_suffix('').name
@@ -56,7 +57,7 @@ class Conll2003DatasetReader(DatasetReader):
                         tags = []
                 else:
                     if self.provide_pos:
-                        token, *_, pos, tag = line.split()
+                        token, pos, *_, tag = line.split()
                         pos_tags.append(pos)
                     else:
                         token, *_, tag = line.split()
