+++ deeppavlov/models/ner/network.py
@@ -145,10 +145,10 @@ class NerNetwork(TFModel):
         if net_type == 'rnn':
             if use_cudnn_rnn:
                 if l2_reg > 0:
-                    raise Warning('cuDNN RNN are not l2 regularizable')
+                    log.warning('cuDNN RNN are not l2 regularizable')
                 units = self._build_cudnn_rnn(features, n_hidden_list, cell_type, intra_layer_dropout, self.mask_ph)
             else:
-                units = self._build_rnn(features, n_hidden_list, cell_type, intra_layer_dropout, self.mask_ph,)
+                units = self._build_rnn(features, n_hidden_list, cell_type, intra_layer_dropout, self.mask_ph)
         elif net_type == 'cnn':
             units = self._build_cnn(features, n_hidden_list, cnn_filter_width, use_batch_norm)
         self._logits = self._build_top(units, n_tags, n_hidden_list[-1], top_dropout, two_dense_on_top)
@@ -163,7 +163,7 @@ class NerNetwork(TFModel):
         sess_config.gpu_options.allow_growth = True
         if gpu is not None:
             sess_config.gpu_options.visible_device_list = str(gpu)
-        self.sess = tf.Session()   # TODO: add sess_config
+        self.sess = tf.Session(config=sess_config)
         self.sess.run(tf.global_variables_initializer())
         super().__init__(**kwargs)
         self.load()
@@ -225,9 +225,11 @@ class NerNetwork(TFModel):
                     units = variational_dropout(units, self._dropout_ph)
             return units
 
-    def _build_rnn(self, units, n_hidden_list, cell_type, intra_layer_dropout):
+    def _build_rnn(self, units, n_hidden_list, cell_type, intra_layer_dropout, mask):
+        sequence_lengths = tf.to_int32(tf.reduce_sum(mask, axis=1))
         for n, n_hidden in enumerate(n_hidden_list):
-            units, _ = bi_rnn(units, n_hidden, cell_type=cell_type, name='Layer_' + str(n))
+            units, _ = bi_rnn(units, n_hidden, cell_type=cell_type,
+                              seq_lengths=sequence_lengths, name='Layer_' + str(n))
             units = tf.concat(units, -1)
             if intra_layer_dropout and n != len(n_hidden_list) - 1:
                 units = variational_dropout(units, self._dropout_ph)
@@ -320,7 +322,8 @@ class NerNetwork(TFModel):
     def train_on_batch(self, *args):
         *xs, y = args
         feed_dict = self._fill_feed_dict(xs, y, train=True, learning_rate=self._learning_rate)
-        self.sess.run(self.train_op, feed_dict)
+        _, loss = self.sess.run([self.train_op, self.loss], feed_dict)
+        return loss
 
     def process_event(self, event_name, data):
         if event_name == 'after_validation':
